Create Gemini Payout Report (#2859)
