+++ app/controllers/admin/publishers_controller.rb
@@ -110,7 +110,7 @@ class Admin::PublishersController < AdminController
   def refresh_uphold
     connection = UpholdConnection.find_by(publisher: params[:publisher_id])
     if connection.present?
-      connection.sync_from_uphold!
+      connection.sync_connection!
       connection.create_uphold_cards
     end
     redirect_to admin_publisher_path(@publisher.id)
