+++ app/models/potential_payment.rb
@@ -4,7 +4,7 @@ class PotentialPayment < ApplicationRecord
   MANUAL = "manual".freeze
 
   # valid wallet_providers
-  enum wallet_provider: { uphold: 0, paypal: 1 }
+  enum wallet_provider: { uphold: 0, paypal: 1, gemini: 2 }
 
   belongs_to :payout_report
   belongs_to :publisher
@@ -15,10 +15,17 @@ class PotentialPayment < ApplicationRecord
   validate :channel_id_not_present_for_referral_payment, if: -> { kind == REFERRAL }
   validate :publisher_id_unique_for_referral_payments
 
-  validates_inclusion_of :reauthorization_needed, :suspended, :uphold_member, :in => [true, false], unless: -> { wallet_provider == 'paypal' }
+  validates_inclusion_of :reauthorization_needed, :suspended, :uphold_member, :in => [true, false], unless: -> { wallet_provider == 'paypal' || wallet_provider == 'gemini' }
+
+  scope :uphold_kyc, -> {
+    where(uphold_status: "ok", reauthorization_needed: false, uphold_member: true, suspended: false)
+  }
+  scope :gemini_kyc, -> {
+    where(gemini_is_verified: true)
+  }
 
   scope :to_be_paid, -> {
-    where(uphold_status: "ok", reauthorization_needed: false, uphold_member: true, suspended: false).
+    uphold_kyc.or(gemini_kyc).
       where("amount::numeric > ?", 0).
       where.not(address: "").
       where.not(address: nil)
