+++ db/migrate/20200827225818_add_columns_to_potential_payment.rb
@@ -1,0 +1,5 @@
+class AddColumnsToPotentialPayment < ActiveRecord::Migration[6.0]
+  def change
+    add_column :potential_payments, :gemini_is_verified, :boolean, default: false
+  end
+end
