+++ test/fixtures/gemini_connections.yml
@@ -4,7 +4,7 @@ default_connection:
   publisher: default
 
 connection_with_token:
-  publisher: completed
+  publisher: gemini_completed
   encrypted_access_token:  "<%= TotpRegistration.encrypt_secret(
     'access_token',
     key: GeminiConnection.new.send(:encryption_key),
