+++ test/fixtures/publishers.yml
@@ -46,6 +46,14 @@ completed:
   default_currency_confirmed_at: "<%= DateTime.now %>"
   subscribed_to_marketing_emails: true
 
+gemini_completed:
+  email: "gemini@completed.org"
+  name: "Gemini the Completed"
+  two_factor_prompted_at: "<%= 1.day.ago %>"
+  agreed_to_tos: "<%= 1.day.ago %>"
+  default_currency_confirmed_at: "<%= DateTime.now %>"
+  subscribed_to_marketing_emails: true
+
 completed_partner:
   email: "partner@completed.org"
   name: "Paul the Partner"
