+++ test/services/payout/paypal_service_test.rb
@@ -1,7 +1,7 @@
 require "test_helper"
 require "webmock/minitest"
 
-class Payout::PotentialPayment::PaypalServiceTest < ActiveJob::TestCase
+class PaypalServiceTest < ActiveJob::TestCase
   include EyeshadeHelper
 
   before do
@@ -47,7 +47,7 @@ class Payout::PotentialPayment::PaypalServiceTest < ActiveJob::TestCase
 
       let(:subject) do
         perform_enqueued_jobs do
-          Payout::PotentialPayment::PaypalService.new(
+          Payout::PaypalService.new(
             payout_report: PayoutReport.create(expected_num_payments: PayoutReport.expected_num_payments(Publisher.all)),
             publisher: publisher,
             should_send_notifications: should_send_notifications).perform
